Change integration test strategy
Rather than looking for ag at a specific path with specific ownership
and permissions, test that we can execute it and get an expected result.
This is much more in line with what an integration test should be, and
also makes it easier to test scenarios such as binary package installs.